+++ src/main/java/com/dealership/Vehicle.java
@@ -1,11 +1,9 @@
 package com.dealership;
 
-// VV Below me is the Vehicle class (Notes) VV
-// VV This is the blueprint for all cars in the dealership VV
 public class Vehicle {
 
     private String vin;
-    private double year;
+    private int year;
     private String make;
     private String model;
     private String vehicleType;
@@ -13,8 +11,7 @@ public class Vehicle {
     private double odometer;
     private double price;
 
-    // VV Below me builds a Vehicle object (Notes) VV
-    public Vehicle(String vin, double year, String make, String model,
+    public Vehicle(String vin, int year, String make, String model,
                    String vehicleType, String color, double odometer, double price) {
 
         this.vin = vin;
@@ -27,9 +24,8 @@ public class Vehicle {
         this.price = price;
     }
 
-    // VV Below me are the getter methods (Notes) VV
     public String getVin() { return vin; }
-    public double getYear() { return year; }
+    public int getYear() { return year; }
     public String getMake() { return make; }
     public String getModel() { return model; }
     public String getVehicleType() { return vehicleType; }
